refactor(process_tracker): remove `ProcessTrackerExt` trait and move remaining methods to impl blocks
