+++ crates/scheduler/src/consumer.rs
@@ -18,9 +18,8 @@ use uuid::Uuid;
 use super::env::logger;
 pub use super::workflows::ProcessTrackerWorkflow;
 use crate::{
-    configs::settings::SchedulerSettings,
-    db::process_tracker::{ProcessTrackerExt, ProcessTrackerInterface},
-    errors, metrics, utils as pt_utils, SchedulerAppState, SchedulerInterface,
+    configs::settings::SchedulerSettings, db::process_tracker::ProcessTrackerInterface, errors,
+    metrics, utils as pt_utils, SchedulerAppState, SchedulerInterface,
 };
 
 // Valid consumer business statuses
